learntris.py: modified the value of key "t"
In earlier version of `execute_instruction`:
> 't': myGame.active_tetramino.show

This does not work because it attaches only the original value of
myGame.active_tetramino with 't'. So, any updates made to
`myGame.active_tetramino` after `execute_instruction` has been defined
once, will not be reflected in the key 't'.

Hence we need to use a lambda function there:
> 't': (lambda: myGame.active_tetramino.show())